리팩토링: 스타일 분리를 위해 프로젝트 구조 변경 - 2
- 유저 정보를 보여주는 PlayerInfo 컴포넌트에서 스타일을 분리
