+++ src/components/player/PlayerInfo/PlayerRankInfo/index.js
@@ -1,0 +1,3 @@
+import PlayerRankInfo from './PlayerRankInfo';
+
+export default PlayerRankInfo;
