+++ src/components/player/PlayerInfo/index.js
@@ -1,0 +1,3 @@
+import PlayerInfo from './PlayerInfo';
+
+export default PlayerInfo;
